# openapi.yaml
openapi: 3.1.0
info:
  title: AEGIS Scanner
  version: 1.3.0
  description: Semgrep-as-a-service for the AEGIS agent fleet. Guild imports this spec as a custom integration.
servers:
  - url: https://REPLACE-WITH-TUNNEL.trycloudflare.com
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-AEGIS-Key
...
security:
  - apiKey: []
paths:
...
  /second-opinion:
    post:
      operationId: second_opinion
      summary: Scan a code snippet with the local AEGIS rules AND with Semgrep's hosted MCP server (mcp.semgrep.ai, tool semgrep_scan), and return both plus the overlap by line. available=false with a reason when the hosted server is down or needs OAuth (set SEMGREP_MCP_TOKEN); the local result is always returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
                language: { type: string, default: python }
                filename: { type: string }
      responses:
        '200':
          description: local and hosted findings
          content:
            application/json:
              schema:
                type: object
                properties:
                  available: { type: boolean }
                  local: { type: array, items: { type: object } }
                  hosted: { type: object, properties: { available: { type: boolean }, reason: { type: string }, findings: { type: array, items: { type: object } } } }
                  overlap_lines: { type: array, items: { type: integer } }
                  only_local_lines: { type: array, items: { type: integer } }
                  only_hosted_lines: { type: array, items: { type: integer } }
                  ms: { type: integer }